backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
openai>=1.0.0
openpyxl>=3.1.0
